+++ data/image-manifest.js
@@ -1,0 +1,39 @@
+window.IMAGE_MANIFEST = {
+  "sights": {
+    "The Vennel": "./assets/images/sights/the-vennel.svg",
+    "Victoria Terrace": "./assets/images/sights/victoria-terrace.svg",
+    "Dean Village": "./assets/images/sights/dean-village.svg",
+    "Circus Lane": "./assets/images/sights/circus-lane.svg",
+    "Regent Road": "./assets/images/sights/regent-road.svg",
+    "Calton Hill": "./assets/images/sights/calton-hill.svg",
+    "Old Man of Storr": "./assets/images/sights/old-man-of-storr.svg",
+    "Quiraing": "./assets/images/sights/quiraing.svg",
+    "Neist Point": "./assets/images/sights/neist-point.svg",
+    "Talisker Bay": "./assets/images/sights/talisker-bay.svg",
+    "Eilean Donan Castle": "./assets/images/sights/eilean-donan-castle.svg",
+    "Bealach na Bà Viewpoint": "./assets/images/sights/bealach-na-ba-viewpoint.svg",
+    "Applecross": "./assets/images/sights/applecross.svg",
+    "Bealach na Gaoithe Viewpoint": "./assets/images/sights/bealach-na-gaoithe-viewpoint.svg",
+    "Dunnottar Castle": "./assets/images/sights/dunnottar-castle.svg",
+    "Glasgow City Centre": "./assets/images/sights/glasgow-city-centre.svg"
+  },
+  "restaurants": {
+    "Makars Mash Bar": "./assets/images/restaurants/makars-mash-bar.svg",
+    "Ting Thai Caravan": "./assets/images/restaurants/ting-thai-caravan.svg",
+    "Garrison West": "./assets/images/restaurants/garrison-west.svg",
+    "Ben Nevis Inn": "./assets/images/restaurants/ben-nevis-inn.svg",
+    "Edinbane Inn": "./assets/images/restaurants/edinbane-inn.svg",
+    "Sea Breezes": "./assets/images/restaurants/sea-breezes.svg",
+    "The Old School Restaurant": "./assets/images/restaurants/the-old-school-restaurant.svg",
+    "The Three Chimneys at Talisker": "./assets/images/restaurants/the-three-chimneys-at-talisker.svg",
+    "The Oyster Shed": "./assets/images/restaurants/the-oyster-shed.svg",
+    "Applecross Inn": "./assets/images/restaurants/applecross-inn.svg",
+    "Café Andaluz Aberdeen": "./assets/images/restaurants/cafe-andaluz-aberdeen.svg",
+    "The Waterfront Café Bar": "./assets/images/restaurants/the-waterfront-cafe-bar.svg",
+    "The Butterfly and the Pig": "./assets/images/restaurants/the-butterfly-and-the-pig.svg",
+    "Ox and Finch": "./assets/images/restaurants/ox-and-finch.svg",
+    "The Finnieston Bar & Restaurant": "./assets/images/restaurants/the-finnieston-bar-restaurant.svg",
+    "Café Gandolfi": "./assets/images/restaurants/cafe-gandolfi.svg",
+    "Sarti Bath Street": "./assets/images/restaurants/sarti-bath-street.svg"
+  }
+};